fastapi_backend/requirements/prod.txt
bcrypt==4.0.1
fastapi==0.95.0
PyJWT==2.6.0
passlib==1.7.4
peewee==3.16.0
peewee-migrate==1.7.0
python-multipart==0.0.6
uvicorn==0.21.1
